CPack: depend on mosquitto from EPEL
Mosquitto is now available in EPEL, so depend on that in the RPM
package. Update documentation to build against the EPEL version.

Unfortunately mosquitto is not available in EPEL for EL6, and the old
repository we used to get mosquitto from is no longer available.
Therefore EL6 will no longer be supported.